+++ wavesrv/pkg/sstore/dbops.go
@@ -13,7 +13,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
 	"github.com/sawka/txwrap"
 	"github.com/wavetermdev/waveterm/waveshell/pkg/base"
@@ -24,9 +23,6 @@ import (
 	"github.com/wavetermdev/waveterm/wavesrv/pkg/scbase"
 	"github.com/wavetermdev/waveterm/wavesrv/pkg/scbus"
 )
-
-const HistoryCols = "h.historyid, h.ts, h.userid, h.sessionid, h.screenid, h.lineid, h.haderror, h.cmdstr, h.remoteownerid, h.remoteid, h.remotename, h.ismetacmd, h.linenum, h.exitcode, h.durationms, h.festate, h.tags, h.status"
-const DefaultMaxHistoryItems = 1000
 
 var updateWriterCVar = sync.NewCond(&sync.Mutex{})
 var WebScreenPtyPosLock = &sync.Mutex{}
@@ -232,178 +228,6 @@ func UpdateRemoteStateVars(ctx context.Context, remoteId string, stateVars map[s
 		query := `UPDATE remote SET statevars = ? WHERE remoteid = ?`
 		tx.Exec(query, quickJson(stateVars), remoteId)
 		return nil
-	})
-}
-
-func InsertHistoryItem(ctx context.Context, hitem *HistoryItemType) error {
-	if hitem == nil {
-		return fmt.Errorf("cannot insert nil history item")
-	}
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		query := `INSERT INTO history 
-                  ( historyid, ts, userid, sessionid, screenid, lineid, haderror, cmdstr, remoteownerid, remoteid, remotename, ismetacmd, linenum, exitcode, durationms, festate, tags, status) VALUES
-                  (:historyid,:ts,:userid,:sessionid,:screenid,:lineid,:haderror,:cmdstr,:remoteownerid,:remoteid,:remotename,:ismetacmd,:linenum,:exitcode,:durationms,:festate,:tags,:status)`
-		tx.NamedExec(query, hitem.ToMap())
-		return nil
-	})
-	return txErr
-}
-
-const HistoryQueryChunkSize = 1000
-
-func _getNextHistoryItem(items []*HistoryItemType, index int, filterFn func(*HistoryItemType) bool) (*HistoryItemType, int) {
-	for ; index < len(items); index++ {
-		item := items[index]
-		if filterFn(item) {
-			return item, index
-		}
-	}
-	return nil, index
-}
-
-// returns true if done, false if we still need to process more items
-func (result *HistoryQueryResult) processItem(item *HistoryItemType, rawOffset int) bool {
-	if result.prevItems < result.Offset {
-		result.prevItems++
-		return false
-	}
-	if len(result.Items) == result.MaxItems {
-		result.HasMore = true
-		result.NextRawOffset = rawOffset
-		return true
-	}
-	if len(result.Items) == 0 {
-		result.RawOffset = rawOffset
-	}
-	result.Items = append(result.Items, item)
-	return false
-}
-
-func runHistoryQueryWithFilter(tx *TxWrap, opts HistoryQueryOpts) (*HistoryQueryResult, error) {
-	if opts.MaxItems == 0 {
-		return nil, fmt.Errorf("invalid query, maxitems is 0")
-	}
-	rtn := &HistoryQueryResult{Offset: opts.Offset, MaxItems: opts.MaxItems}
-	var rawOffset int
-	if opts.RawOffset >= opts.Offset {
-		rtn.prevItems = opts.Offset
-		rawOffset = opts.RawOffset
-	} else {
-		rawOffset = 0
-	}
-	for {
-		resultItems, err := runHistoryQuery(tx, opts, rawOffset, HistoryQueryChunkSize)
-		if err != nil {
-			return nil, err
-		}
-		isDone := false
-		for resultIdx := 0; resultIdx < len(resultItems); resultIdx++ {
-			if opts.FilterFn != nil && !opts.FilterFn(resultItems[resultIdx]) {
-				continue
-			}
-			isDone = rtn.processItem(resultItems[resultIdx], rawOffset+resultIdx)
-			if isDone {
-				break
-			}
-		}
-		if isDone {
-			break
-		}
-		if len(resultItems) < HistoryQueryChunkSize {
-			break
-		}
-		rawOffset += HistoryQueryChunkSize
-	}
-	return rtn, nil
-}
-
-func runHistoryQuery(tx *TxWrap, opts HistoryQueryOpts, realOffset int, itemLimit int) ([]*HistoryItemType, error) {
-	// check sessionid/screenid format because we are directly inserting them into the SQL
-	if opts.SessionId != "" {
-		_, err := uuid.Parse(opts.SessionId)
-		if err != nil {
-			return nil, fmt.Errorf("malformed sessionid")
-		}
-	}
-	if opts.ScreenId != "" {
-		_, err := uuid.Parse(opts.ScreenId)
-		if err != nil {
-			return nil, fmt.Errorf("malformed screenid")
-		}
-	}
-	if opts.RemoteId != "" {
-		_, err := uuid.Parse(opts.RemoteId)
-		if err != nil {
-			return nil, fmt.Errorf("malformed remoteid")
-		}
-	}
-	whereClause := "WHERE 1"
-	var queryArgs []interface{}
-	hNumStr := ""
-	if opts.SessionId != "" && opts.ScreenId != "" {
-		whereClause += fmt.Sprintf(" AND h.sessionid = '%s' AND h.screenid = '%s'", opts.SessionId, opts.ScreenId)
-		hNumStr = ""
-	} else if opts.SessionId != "" {
-		whereClause += fmt.Sprintf(" AND h.sessionid = '%s'", opts.SessionId)
-		hNumStr = "s"
-	} else {
-		hNumStr = "g"
-	}
-	if opts.SearchText != "" {
-		whereClause += " AND h.cmdstr LIKE ? ESCAPE '\\'"
-		likeArg := opts.SearchText
-		likeArg = strings.ReplaceAll(likeArg, "%", "\\%")
-		likeArg = strings.ReplaceAll(likeArg, "_", "\\_")
-		queryArgs = append(queryArgs, "%"+likeArg+"%")
-	}
-	if opts.FromTs > 0 {
-		whereClause += fmt.Sprintf(" AND h.ts <= %d", opts.FromTs)
-	}
-	if opts.RemoteId != "" {
-		whereClause += fmt.Sprintf(" AND h.remoteid = '%s'", opts.RemoteId)
-	}
-	if opts.NoMeta {
-		whereClause += " AND NOT h.ismetacmd"
-	}
-	query := fmt.Sprintf("SELECT %s, ('%s' || CAST((row_number() OVER win) as text)) historynum FROM history h %s WINDOW win AS (ORDER BY h.ts, h.historyid) ORDER BY h.ts DESC, h.historyid DESC LIMIT %d OFFSET %d", HistoryCols, hNumStr, whereClause, itemLimit, realOffset)
-	marr := tx.SelectMaps(query, queryArgs...)
-	rtn := make([]*HistoryItemType, len(marr))
-	for idx, m := range marr {
-		hitem := dbutil.FromMap[*HistoryItemType](m)
-		rtn[idx] = hitem
-	}
-	return rtn, nil
-}
-
-func GetHistoryItems(ctx context.Context, opts HistoryQueryOpts) (*HistoryQueryResult, error) {
-	var rtn *HistoryQueryResult
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		var err error
-		rtn, err = runHistoryQueryWithFilter(tx, opts)
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	if txErr != nil {
-		return nil, txErr
-	}
-	return rtn, nil
-}
-
-func GetHistoryItemByLineNum(ctx context.Context, screenId string, lineNum int) (*HistoryItemType, error) {
-	return WithTxRtn(ctx, func(tx *TxWrap) (*HistoryItemType, error) {
-		query := `SELECT * FROM history WHERE screenid = ? AND linenum = ?`
-		hitem := dbutil.GetMapGen[*HistoryItemType](tx, query, screenId, lineNum)
-		return hitem, nil
-	})
-}
-
-func GetLastHistoryLineNum(ctx context.Context, screenId string) (int, error) {
-	return WithTxRtn(ctx, func(tx *TxWrap) (int, error) {
-		query := `SELECT COALESCE(max(linenum), 0) FROM history WHERE screenid = ?`
-		maxLineNum := tx.GetInt(query, screenId)
-		return maxLineNum, nil
 	})
 }
 
@@ -2335,194 +2159,6 @@ func GetDBVersion(ctx context.Context) (int, error) {
 		return nil
 	})
 	return version, txErr
-}
-
-type bookmarkOrderType struct {
-	BookmarkId string
-	OrderIdx   int64
-}
-
-func GetBookmarks(ctx context.Context, tag string) ([]*BookmarkType, error) {
-	var bms []*BookmarkType
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		var query string
-		if tag == "" {
-			query = `SELECT * FROM bookmark`
-			bms = dbutil.SelectMapsGen[*BookmarkType](tx, query)
-		} else {
-			query = `SELECT * FROM bookmark WHERE EXISTS (SELECT 1 FROM json_each(tags) WHERE value = ?)`
-			bms = dbutil.SelectMapsGen[*BookmarkType](tx, query, tag)
-		}
-		bmMap := dbutil.MakeGenMap(bms)
-		var orders []bookmarkOrderType
-		query = `SELECT bookmarkid, orderidx FROM bookmark_order WHERE tag = ?`
-		tx.Select(&orders, query, tag)
-		for _, bmOrder := range orders {
-			bm := bmMap[bmOrder.BookmarkId]
-			if bm != nil {
-				bm.OrderIdx = bmOrder.OrderIdx
-			}
-		}
-		return nil
-	})
-	if txErr != nil {
-		return nil, txErr
-	}
-	return bms, nil
-}
-
-func GetBookmarkById(ctx context.Context, bookmarkId string, tag string) (*BookmarkType, error) {
-	var rtn *BookmarkType
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		query := `SELECT * FROM bookmark WHERE bookmarkid = ?`
-		rtn = dbutil.GetMapGen[*BookmarkType](tx, query, bookmarkId)
-		if rtn == nil {
-			return nil
-		}
-		query = `SELECT orderidx FROM bookmark_order WHERE bookmarkid = ? AND tag = ?`
-		orderIdx := tx.GetInt(query, bookmarkId, tag)
-		rtn.OrderIdx = int64(orderIdx)
-		return nil
-	})
-	if txErr != nil {
-		return nil, txErr
-	}
-	return rtn, nil
-}
-
-func GetBookmarkIdByArg(ctx context.Context, bookmarkArg string) (string, error) {
-	var rtnId string
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		if len(bookmarkArg) == 8 {
-			query := `SELECT bookmarkid FROM bookmark WHERE bookmarkid LIKE (? || '%')`
-			rtnId = tx.GetString(query, bookmarkArg)
-			return nil
-		}
-		query := `SELECT bookmarkid FROM bookmark WHERE bookmarkid = ?`
-		rtnId = tx.GetString(query, bookmarkArg)
-		return nil
-	})
-	if txErr != nil {
-		return "", txErr
-	}
-	return rtnId, nil
-}
-
-func GetBookmarkIdsByCmdStr(ctx context.Context, cmdStr string) ([]string, error) {
-	return WithTxRtn(ctx, func(tx *TxWrap) ([]string, error) {
-		query := `SELECT bookmarkid FROM bookmark WHERE cmdstr = ?`
-		bmIds := tx.SelectStrings(query, cmdStr)
-		return bmIds, nil
-	})
-}
-
-// ignores OrderIdx field
-func InsertBookmark(ctx context.Context, bm *BookmarkType) error {
-	if bm == nil || bm.BookmarkId == "" {
-		return fmt.Errorf("invalid empty bookmark id")
-	}
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		query := `SELECT bookmarkid FROM bookmark WHERE bookmarkid = ?`
-		if tx.Exists(query, bm.BookmarkId) {
-			return fmt.Errorf("bookmarkid already exists")
-		}
-		query = `INSERT INTO bookmark ( bookmarkid, createdts, cmdstr, alias, tags, description)
-                               VALUES (:bookmarkid,:createdts,:cmdstr,:alias,:tags,:description)`
-		tx.NamedExec(query, bm.ToMap())
-		for _, tag := range append(bm.Tags, "") {
-			query = `SELECT COALESCE(max(orderidx), 0) FROM bookmark_order WHERE tag = ?`
-			maxOrder := tx.GetInt(query, tag)
-			query = `INSERT INTO bookmark_order (tag, bookmarkid, orderidx) VALUES (?, ?, ?)`
-			tx.Exec(query, tag, bm.BookmarkId, maxOrder+1)
-		}
-		return nil
-	})
-	return txErr
-}
-
-const (
-	BookmarkField_Desc   = "desc"
-	BookmarkField_CmdStr = "cmdstr"
-)
-
-func EditBookmark(ctx context.Context, bookmarkId string, editMap map[string]interface{}) error {
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		query := `SELECT bookmarkid FROM bookmark WHERE bookmarkid = ?`
-		if !tx.Exists(query, bookmarkId) {
-			return fmt.Errorf("bookmark not found")
-		}
-		if desc, found := editMap[BookmarkField_Desc]; found {
-			query = `UPDATE bookmark SET description = ? WHERE bookmarkid = ?`
-			tx.Exec(query, desc, bookmarkId)
-		}
-		if cmdStr, found := editMap[BookmarkField_CmdStr]; found {
-			query = `UPDATE bookmark SET cmdstr = ? WHERE bookmarkid = ?`
-			tx.Exec(query, cmdStr, bookmarkId)
-		}
-		return nil
-	})
-	return txErr
-}
-
-func fixupBookmarkOrder(tx *TxWrap) {
-	query := `
-WITH new_order AS (
-  SELECT tag, bookmarkid, row_number() OVER (PARTITION BY tag ORDER BY orderidx) AS newidx FROM bookmark_order
-)
-UPDATE bookmark_order
-SET orderidx = new_order.newidx
-FROM new_order
-WHERE bookmark_order.tag = new_order.tag AND bookmark_order.bookmarkid = new_order.bookmarkid
-`
-	tx.Exec(query)
-}
-
-func DeleteBookmark(ctx context.Context, bookmarkId string) error {
-	txErr := WithTx(ctx, func(tx *TxWrap) error {
-		query := `SELECT bookmarkid FROM bookmark WHERE bookmarkid = ?`
-		if !tx.Exists(query, bookmarkId) {
-			return fmt.Errorf("bookmark not found")
-		}
-		query = `DELETE FROM bookmark WHERE bookmarkid = ?`
-		tx.Exec(query, bookmarkId)
-		query = `DELETE FROM bookmark_order WHERE bookmarkid = ?`
-		tx.Exec(query, bookmarkId)
-		fixupBookmarkOrder(tx)
-		return nil
-	})
-	return txErr
-}
-
-func getLineIdsFromHistoryItems(historyItems []*HistoryItemType) []string {
-	var rtn []string
-	for _, hitem := range historyItems {
-		if hitem.LineId != "" {
-			rtn = append(rtn, hitem.LineId)
-		}
-	}
-	return rtn
-}
-
-func GetLineCmdsFromHistoryItems(ctx context.Context, historyItems []*HistoryItemType) ([]*LineType, []*CmdType, error) {
-	if len(historyItems) == 0 {
-		return nil, nil, nil
-	}
-	return WithTxRtn3(ctx, func(tx *TxWrap) ([]*LineType, []*CmdType, error) {
-		lineIdsJsonArr := quickJsonArr(getLineIdsFromHistoryItems(historyItems))
-		query := `SELECT * FROM line WHERE lineid IN (SELECT value FROM json_each(?))`
-		lineArr := dbutil.SelectMappable[*LineType](tx, query, lineIdsJsonArr)
-		query = `SELECT * FROM cmd WHERE lineid IN (SELECT value FROM json_each(?))`
-		cmdArr := dbutil.SelectMapsGen[*CmdType](tx, query, lineIdsJsonArr)
-		return lineArr, cmdArr, nil
-	})
-}
-
-func PurgeHistoryByIds(ctx context.Context, historyIds []string) error {
-	return WithTx(ctx, func(tx *TxWrap) error {
-		query := `DELETE FROM history WHERE historyid IN (SELECT value FROM json_each(?))`
-		tx.Exec(query, quickJsonArr(historyIds))
-		return nil
-	})
 }
 
 func CountScreenWebShares(ctx context.Context) (int, error) {
