+++ wavesrv/pkg/sstore/updatetypes.go
@@ -91,18 +91,6 @@ func (ClearInfoUpdate) GetType() string {
 	return "clearinfo"
 }
 
-type HistoryInfoType struct {
-	HistoryType string             `json:"historytype"`
-	SessionId   string             `json:"sessionid,omitempty"`
-	ScreenId    string             `json:"screenid,omitempty"`
-	Items       []*HistoryItemType `json:"items"`
-	Show        bool               `json:"show"`
-}
-
-func (HistoryInfoType) GetType() string {
-	return "history"
-}
-
 type InteractiveUpdate bool
 
 func (InteractiveUpdate) GetType() string {
@@ -120,43 +108,6 @@ type ConnectUpdate struct {
 
 func (ConnectUpdate) GetType() string {
 	return "connect"
-}
-
-type MainViewUpdate struct {
-	MainView      string           `json:"mainview"`
-	HistoryView   *HistoryViewData `json:"historyview,omitempty"`
-	BookmarksView *BookmarksUpdate `json:"bookmarksview,omitempty"`
-}
-
-func (MainViewUpdate) GetType() string {
-	return "mainview"
-}
-
-type BookmarksUpdate struct {
-	Bookmarks        []*BookmarkType `json:"bookmarks"`
-	SelectedBookmark string          `json:"selectedbookmark,omitempty"`
-}
-
-func (BookmarksUpdate) GetType() string {
-	return "bookmarks"
-}
-
-func AddBookmarksUpdate(update *scbus.ModelUpdatePacketType, bookmarks []*BookmarkType, selectedBookmark *string) {
-	if selectedBookmark == nil {
-		update.AddUpdate(BookmarksUpdate{Bookmarks: bookmarks})
-	} else {
-		update.AddUpdate(BookmarksUpdate{Bookmarks: bookmarks, SelectedBookmark: *selectedBookmark})
-	}
-}
-
-type HistoryViewData struct {
-	Items         []*HistoryItemType `json:"items"`
-	Offset        int                `json:"offset"`
-	RawOffset     int                `json:"rawoffset"`
-	NextRawOffset int                `json:"nextrawoffset"`
-	HasMore       bool               `json:"hasmore"`
-	Lines         []*LineType        `json:"lines"`
-	Cmds          []*CmdType         `json:"cmds"`
 }
 
 type RemoteEditType struct {
